requirements.txt
django==5.2
pillow==12.0.0
django-cleanup==9.0.0
django-bootstrap5==25.3
django-autoslug==1.9.9